% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bReakingbad.R
\docType{package}
\name{bReakingbad}
\alias{bReakingbad}
\title{bReakingbad: a package for formatting of p-values in markdown tables}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Formatting functions}{

The foo functions ...
}

