% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFormatters.R
\name{to_proper_case}
\alias{to_proper_case}
\title{Proper case}
\usage{
to_proper_case(x)
}
\arguments{
\item{x}{the string to be formatted}
}
\description{
Converts 1st character of string to upper case and all other characters to lower case
}
