% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFormatters.R
\name{to_proper_case}
\alias{to_proper_case}
\title{Proper case}
\usage{
to_proper_case(x)
}
\arguments{
\item{x}{the string to be formatted}
}
\value{
the string in proper case
}
\description{
Converts 1st character of string to upper case and all other characters to lower case. Refer to https://stackoverflow.com/questions/24956546/capitalizing-letters-r-equivalent-of-excel-proper-function
}
\examples{
to_proper_case(x = "An Example String")
}
